+++ src/api/type-defs/inputs/SetTimestampOffsetByFilterTaskInput.ts
@@ -1,6 +1,6 @@
 export default /* GraphQL */ `
   input SetTimestampOffsetByFilterTaskInput {
     filters: FiltersInput!
-    offsetMs: Int!
+    offsetMs: Float!
   }
 `;
